requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.14,<4.0
django-oauth-toolkit==3.0.1
django-allauth
dj-rest-auth
social-auth-app-django
pymysql
python-dotenv
drf-yasg
social_django
rest_framework_simplejwt
celery
redis


redis-server
 celery -A authApi  worker --loglevel=info